\docType{package}
\name{capm-package}
\alias{capm}
\alias{capm-package}
\title{The capm Package}
\description{
  Companion Animal Population Management.
}
\details{
  \tabular{ll}{ Package: \tab capm\cr Type: \tab Package\cr
  Version: \tab 0.1.0\cr Date: \tab 2014-01-28\cr Depends:
  \tab R (>= 2.15.0), FME, survey, reshape2, ggplot2, grid,
  rgdal, maptools\cr Encoding: \tab UTF-8\cr License: \tab
  GPL (>= 2)\cr LazyLoad: \tab yes\cr URL: \tab
  \url{https://github.com/oswaldosantos/capm}\cr }

  Provides functions for quantitative companion animal
  population management. Further information can be found
  in the link given above.
}
\author{
  Creator: Oswaldo Santos \email{oswaldosant@gmail.com}\cr
  Maintainer: Oswaldo Santos
  \email{oswaldosant@gmail.com}\cr Contributors: Fernando
  Silveira Marques \email{fernandosix@gmail.com}, Marcos
  Amaku \email{amaku@vps.fmvz.usp.br}, Fernando Ferreira
  \email{fernando@vps.fmvz.usp.br}
}
\keyword{package}

