% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\docType{data}
\name{city}
\alias{city}
\title{Census tracks of Santos, Brazil.}
\format{A data frame with 655 observations and 5 variables:
\describe{
  \item{track_id}{Census track's ID's of Santos, Brazil.}
  \item{track_status}{Censu track status. If less than 4, urban, otherwise rural.}
  \item{persons_mean}{Mean number of persons per household.}
  \item{hh}{Number of households.}
  \item{persons}{Number of persons.}
}}
\source{
\url{http://www.ibge.gov.br/}
}
\usage{
city
}
\description{
Census tracks of Santos, Brazil, according to a census of 2010.
}
\keyword{datasets}
