% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_pilot.R
\docType{data}
\name{cluster_pilot}
\alias{cluster_pilot}
\title{Two-stage cluster sample: pilot study}
\format{A data frame with 655 observations and 5 variables:
\describe{
  \item{track_id}{Census track ID of sampled households from Santos, Brazil.}
  \item{dogs}{Number of dogs per household (simulated).}
}}
\source{
\url{http://www.ibge.gov.br/}
}
\usage{
cluster_pilot
}
\description{
Two-stage cluster sample to the calculate sample size of a two-stage cluster design.
}
\keyword{datasets}
