% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sample.R
\docType{data}
\name{cluster_sample}
\alias{cluster_sample}
\title{Household data from a two-stage cluster sample}
\format{Data frame with 1230 observations (households) and 14 variables:
\describe{
  \item{interview_id}{Interview's ID.}
  \item{census_tract_id}{Census tract's ID.}
  \item{interviewer}{Interviewer's name.}
  \item{date}{Interview's date.}
  \item{address}{Household's address.}
  \item{interview}{Interview status.}
  \item{interviewee}{Interviewee's name.}
  \item{number_of_dogs}{Number of dogs.}
  \item{number_of_cats}{Number of cats.}
  \item{number_of_persons}{Number of persons.}
  \item{cell_phone}{Interviewee's cell phone.}
  \item{e_mail}{Interviewee's e-mail.}
  \item{reasons_for_not_sterilize}{Interviewee's reported reason for not sterilizing her/his animal.}
  \item{reasons_for_not_sterilize_others}{Other reasons for not sterilizing her/his animal.}
  }}
\usage{
cluster_sample
}
\description{
Data described and analyzed by Baquero et al., 2018.
}
\references{
Baquero, O. S., Marconcin, S., Rocha, A., & Garcia, R. D. C. M. (2018). Companion animal demography and population management in Pinhais, Brazil. Preventive Veterinary Medicine.
}
\keyword{datasets}
