% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sample_animals_lost.R
\docType{data}
\name{cluster_sample_animals_lost}
\alias{cluster_sample_animals_lost}
\title{Lost owned animals data from a two-stage cluster sample}
\format{A data frame with 110 observations and 8 variables:
#' \describe{
  \item{track_id}{Census track ID from Santos, Brazil.}
  \item{hh_id}{Household ID.}
  \item{name}{Animal's name.}
  \item{species}{Animal's species.}
  \item{sex}{Animal's sex.}
  \item{age}{Animal's age. An age equal to 0 means that the animal had less than 1 year.}
  \item{sterilized}{Animal's reproductive status.}
  \item{fate}{Animal's fate.}
}}
\usage{
cluster_sample_animals_lost
}
\description{
Lost owned animals hypothetical data from a two-stage cluster sammple.
}
\keyword{datasets}
