% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqTab.R
\name{FreqTab}
\alias{FreqTab}
\title{Frequency table of categorical variables}
\usage{
FreqTab(variable = NULL, rnd = 3, decreasing = TRUE, ...)
}
\arguments{
\item{variable}{a categorical variable.}

\item{rnd}{the number of decimal places (round) or significant digits (signif) to be used.}

\item{decreasing}{\code{\link{logical}} (default). If \code{TRUE}, frequencies will be sorted in decreasing order, if \code{FALSE}, they will be sorted in increasing order.}
}
\value{
\code{\link{data.frame}}.
}
\description{
Calculates and sort the count and relative frequency of categories.
}
\examples{
data(hh)
FreqTab(hh$common_fates)

}
\seealso{
\link[base]{table} and \link[base]{sort}.
}
