% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicInterface.R
\name{GraphicInterface}
\alias{GraphicInterface}
\title{Graphic interface to use some capm functions}
\usage{
GraphicInterface(set.func)
}
\arguments{
\item{set.func}{string to select a graphic interface for a set of functions to achieve a specific task. \code{SelectSamplingUnits} creates a graphic interface to select sampling units for pilot or final survey designs. \code{CalculateSampleSize} creates a graphic interface to calculate sample size and composition. The graphic interface created by \code{SurveyAnalysis} support functionality to analyise survey data. \code{SolveIASA} creates an interface to simulate population dynamics and to assess population-based interventions.}
}
\value{
a graphic interface in a browser.
}
\description{
Graphic interface to use some capm functions
}
\details{
The graphic interfaces are created with the \code{shiny} package and thus will open in a browser.
}
\examples{
\dontrun{
GraphicInterface(set.func = 'SelectSamplingUnits')
}
}
\references{
\url{http://oswaldosantos.github.io/capm}
}
