% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.R
\docType{data}
\name{hh}
\alias{hh}
\title{Household data}
\format{A data frame with 146093 observations and 12 variables:
\describe{
  \item{track_id}{Census track's ID from Santos, Brazil.}
  \item{hh_id}{Household's ID.}
  \item{persons}{Number of persons.}
  \item{dogs}{Number of dogs.}
  \item{cats}{Number of cats.}
  \item{reasons_to_not_sterilize}{Reasons the interviewee has to not sterilize her/his animal.}
  #'   \item{reasons_to_not_sterilize_others}{Other reasons the interviewee has to not sterilize her/his animal.}
  \item{common_fates}{Most common fate of animals among selected options, according to the interviewee.}
  #'   \item{common_fates_others}{Other options for the most common fate of animals, according to the interviewee.}
  \item{tolerance}{Responses of owners when their animals destroy an expensive thing, according to the interviewee.}
  \item{tolerance_others}{Other  responses of owners when their animals destroy an expensive thing, according to the interviewee.}
  \item{reasons_to_abandom}{Reasons that would lead the interviewees to abandom their animals.}
}}
\usage{
hh
}
\description{
Household hypothetical data.
}
\keyword{datasets}
