\name{capm-package}
\alias{capm-package}
\alias{capm}
\docType{package}
\title{
Customer Analytics Using Probability Models
}
\description{
Fit latent variable models to understand who your customers are and what they will do next. Contains maximum likelihood estimation routines for a variety of models, as well as S3 methods to explore model fit and produce forecasts. 
}
\details{
\tabular{ll}{
Package: \tab capm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-17\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Eli Stein (eli.manfred.stein@gmail.com)

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
