% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psu_ssu.R
\docType{data}
\name{psu_ssu}
\alias{psu_ssu}
\title{Census tracts of Pinhais, Brazil.}
\format{A data frame with 655 observations and 5 variables:
\describe{
  \item{census_tract_id}{Census track's ID.}
  \item{hh}{Number of households.}
}}
\source{
\url{http://www.ibge.gov.br/}
}
\usage{
psu_ssu
}
\description{
Census tracts of Pinhais, Brazil, according to the census of 2010. Data described and analyzed by Baquero et al., 2018.
}
\references{
Baquero, O. S., Marconcin, S., Rocha, A., & Garcia, R. D. C. M. (2018). Companion animal demography and population management in Pinhais, Brazil. Preventive Veterinary Medicine.
}
\keyword{datasets}
