% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_sample_animals_lost.R
\docType{data}
\name{sys_sample_animals_lost}
\alias{sys_sample_animals_lost}
\title{Lost owned animals data from a two-stage cluster sample}
\format{A data frame with 98 observations and 6 variables:
#' \describe{
  \item{hh_id}{Household ID.}
  \item{name}{Animal's name.}
  \item{species}{Animal's species.}
  \item{sex}{Animal's sex.}
  \item{age}{Animal's age. An age equal to 0 means that the  has less than 1 year.}
  \item{sterilized}{Animal's reproductive status.}
  \item{fate}{Animal's fate.}
}}
\usage{
sys_sample_animals_lost
}
\description{
Lost owned animals data hypothetical data from a two-stage cluster sammple.
}
\keyword{datasets}
