% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm-package.R
\docType{package}
\name{capm-package}
\alias{capm-package}
\title{The capm Package}
\description{
Companion Animal Population Management. Provides functions for quantitative Companion Animal Population Management. Further information can be found in the URL given below.
}
\details{
\tabular{ll}{
Package: \tab capm\cr
Type: \tab Package\cr
Version: \tab 0.12.13\cr
Date: \tab 2018-08-30\cr
Depends: \tab R (>= 3.4.1)\cr
Imports: \tab deSolve, FME, survey, dplyr, tidyr, ggplot2, shiny, grid\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{http://oswaldosantos.github.io/capm}\cr
Author: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
Contributors: \tab Marcos Amaku \email{amaku@vps.fmvz.usp.br}, Fernando Ferreira \email{fernando@vps.fmvz.usp.br}
}
}
\keyword{package}
