\docType{package}
\name{capm-package}
\alias{capm-package}
\title{The capm Package}
\description{
  Companion Animal Population Management.
}
\details{
  \tabular{ll}{ Package: \tab capm\cr Type: \tab Package\cr
  Version: \tab 0.2.5\cr Date: \tab 2014-02-16\cr Depends:
  \tab R (>= 3.0.2), FME, survey, reshape2, ggplot2, shiny,
  grid, rgdal, maptools\cr License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr URL: \tab
  \url{http://oswaldosantos.github.io/capm}\cr }

  Provides functions for quantitative companion animal
  population management. Further information can be found
  in the link given above.
}
\author{
  Author: Oswaldo Santos \email{oswaldosant@gmail.com}\cr
  Maintainer: Oswaldo Santos
  \email{oswaldosant@gmail.com}\cr Contributors: Fernando
  Silveira Marques \email{fernandosix@gmail.com}, Marcos
  Amaku \email{amaku@vps.fmvz.usp.br}, Fernando Ferreira
  \email{fernando@vps.fmvz.usp.br}
}
\keyword{package}

