\docType{package}
\name{capm-package}
\alias{capm}
\alias{capm-package}
\title{The capm Package}
\description{
  Companion Animal Population Management.
}
\details{
  \tabular{ll}{ Package: \tab capm\cr Type: \tab Package\cr
  Version: \tab 0.1.0\cr Date: \tab 2014-01-05\cr Depends:
  \tab R (>= 2.15.0), FME, survey, ggplot2, grid, rgdal,
  maptools\cr Encoding: \tab UTF-8\cr License: \tab GPL (>=
  2)\cr LazyLoad: \tab yes\cr URL: \tab
  https://github.com/oswaldosantos/capm\cr }

  Provides functions to calculate the size and compostion
  of samples to estimate dog and cat demographic parameters
  and functions to simulate population dynamics under
  different scenarios of population management.
}
\author{
  Oswaldo Santos \email{oswaldosant@gmail.com}, Fernando
  Silveira Marques \email{fernandosix@gmail.com}, Marcos
  Amaku \email{amaku@vps.fmvz.usp.br}, Fernando Ferreira
  \email{fernando@vps.fmvz.usp.br}
}
\keyword{package}

