% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm-package.R
\docType{package}
\name{capm-package}
\alias{capm-package}
\title{The capm Package}
\description{
Companion Animal Population Management. Provides functions for quantitative Companion Animal Population Management. Further information can be found in the URL given below.
}
\details{
\tabular{ll}{
Package: \tab capm\cr
Type: \tab Package\cr
Version: \tab 0.7.0\cr
Date: \tab 2015-04-03\cr
Depends: \tab R (>= 3.1.2)\cr
Imports: \tab deSolve, FME, survey, reshape2, ggplot2, shiny, grid, rgdal, maptools, sp\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{http://oswaldosantos.github.io/capm}\cr
Author: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
Contributors: \tab Marcos Amaku \email{amaku@vps.fmvz.usp.br}, Fernando Ferreira \email{fernando@vps.fmvz.usp.br}
}
}
\keyword{package}

