\docType{package}
\name{capm-package}
\alias{capm-package}
\title{The capm Package}
\description{
Companion Animal Population Management. Provides functions
for quantitative Companion Animal Population Management.
Further information can be found in the URL given below.
}
\details{
\tabular{ll}{ Package: \tab capm\cr Type: \tab Package\cr
Version: \tab 0.5\cr Date: \tab 2014-05-27\cr Depends: \tab
R (>= 3.0.2)\cr Imports: \tab deSolve, FME, survey,
reshape2, ggplot2, shiny, grid, rgdal, maptools, sp\cr
License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr URL: \tab
\url{http://oswaldosantos.github.io/capm}\cr Author: \tab
Oswaldo Santos \email{oswaldosant@gmail.com}\cr Maintainer:
\tab Oswaldo Santos \email{oswaldosant@gmail.com}\cr
Contributors: \tab Marcos Amaku
\email{amaku@vps.fmvz.usp.br}, Fernando Ferreira
\email{fernando@vps.fmvz.usp.br} }
}
\keyword{package}

