% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqTab.R
\name{FreqTab}
\alias{FreqTab}
\title{Frequency table of categorical variables}
\usage{
FreqTab(data = NULL, variables = NULL, rnd = 3, decreasing = TRUE, ...)
}
\arguments{
\item{variables}{name or position of categorical variables. If more than one variable is provided, contingency frequencies are calculated.}

\item{rnd}{the number of decimal places (round) or significant digits (signif) to be used.}

\item{decreasing}{\code{\link{logical}} (default). If \code{TRUE}, frequencies will be sorted in decreasing order, if \code{FALSE}, they will be sorted in increasing order.}

\item{\code{\link{data.frame}}}{with categorical variables.}
}
\value{
\code{\link{data.frame}}.
}
\description{
Calculates and sort the count and relative frequency of categories.
}
\examples{
data(hh)
FreqTab(hh, c("common_fates", "tolerance"))

}
\seealso{
\link[base]{table} and \link[base]{sort}.
}
